+++ cli/CHANGELOG.md
@@ -1,5 +1,5 @@
-Unreleased
-----------
+0.1.6
+-----
 - Added `--procmap-query` option to `normalize user` sub-command
 - Bumped `blazesym` dependency to `0.2.0-rc.1`